+++ backend/Repository/CheckoutRepository.cs
@@ -13,6 +13,6 @@ public class CheckoutRepository : BaseRepository<CheckoutLog>
         => await _dbSet.AnyAsync() ? await _dbSet.MaxAsync(l => l.LogId) + 1L : 1L;
 
     /// <summary>某分配进行中的清算单（UK_D_CHECKOUT_ACTIVE 的代码侧快查，数据库约束为最终兜底）</summary>
-    public Task<CheckoutLog?> GetActiveByAllocationAsync(int allocationId)
+    public Task<CheckoutLog?> GetActiveByAllocationAsync(long allocationId)
         => _dbSet.FirstOrDefaultAsync(l => l.AllocationId == allocationId && l.Status == "待清算");
 }
